Patch coverage for fixVsCodeMultiSelect
