--- packages/zowe-explorer-api/src/tree/TreeViewUtils.ts
@@ -1,14 +1,0 @@
-import { IZoweTree } from "./IZoweTree";
-
-export class TreeViewUtils {
-    /**
-     * Temporary solution to fixing issue with VSCode multi-select bug
-     * (where multi-select is still active after deleting the nodes in previous multi-selection)
-     * @param treeProvider The tree provider to reset multi-selection for
-     */
-    public static async fixVsCodeMultiSelect<T>(treeProvider: IZoweTree<T>): Promise<void> {
-        const treeView = treeProvider.getTreeView();
-        await treeView.reveal(treeProvider.mFavoriteSession, { select: true });
-        await treeView.reveal(treeProvider.mFavoriteSession, { select: false });
-    }
-}
